backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sendgrid==6.10.0
python-dotenv==1.0.0
email-validator==2.1.0
pydantic[email]==2.5.0